djapp/requirements.txt
django<3.3
django-filter
django-netfields
djangorestframework
drf_yasg
gunicorn
httptools
psycopg2-binary
python-dotenv
requests
uvicorn
uvloop
openstacksdk
ncclient

oslo.config>=6.0.0 # Apache-2.0
oslo.messaging>=8.0.0 # Apache-2.0

keystoneauth1>=3.16.0 # Apache-2.0
# python-keystoneclient>=4.1.1 # Apache-2.0
python-cinderclient!=4.0.0,>=3.3.0 # Apache-2.0
python-neutronclient>=6.7.0 # Apache-2.0
python-glanceclient>=2.8.0 # Apache-2.0
python-novaclient>=17.2.1 # Apache-2.0


# celery
celery
django-celery-results
# django-celery-beat
